예외처리 1차 처리

+++ src/pages/ObjectionDetail.jsx
@@ -19,8 +19,10 @@ const ObjectionDetail = () => {
 
   //게시글 삭제
   const onDeleteHandler = (payload) => {
-    dispatch(__deleteObjection(payload));
-    window.location.replace("/objectionread/all");
+    if (window.confirm("정말 삭제하시겠습니까?")) {
+      dispatch(__deleteObjection(payload));
+      window.location.replace("/objectionread/all");
+    }
   };
 
   return (
